gemini/pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = ""
authors = [
    {name = "youry07",email = "youry@ukr.net"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai"
]

[project.scripts]
probe = "gemini.probe:main"


[tool.poetry]
packages = [{include = "gemini", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
